[JSC] Share OMGIRGenerator class def between 64/32 bits https://bugs.webkit.org/show_bug.cgi?id=300132
Reviewed by NOBODY (OOPS!).

We're at the point where 32-bits can share the class definition of
OMGIRGenerator with 64-bits, modulo some helpers.

Stop building WasmOMGIRGenerator32_64.cpp separately. Instead, only
build WasmOMGIRGenerator.cpp and #include the 32-bit function
definitions from WasmOMGIRGenerator32_64.cpp.
