+++ Source/JavaScriptCore/wasm/WasmOMGIRGenerator32_64.cpp
@@ -23,1058 +23,9 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "config.h"
-#include "WasmOMGIRGenerator.h"
-
-#if ENABLE(WEBASSEMBLY_OMGJIT)
-
-#include "AirCode.h"
-#include "AllowMacroScratchRegisterUsageIf.h"
-#include "B3AbstractHeapRepository.h"
-#include "B3BasicBlockInlines.h"
-#include "B3CCallValue.h"
-#include "B3Const128Value.h"
-#include "B3ConstPtrValue.h"
-#include "B3EstimateStaticExecutionCounts.h"
-#include "B3FixSSA.h"
-#include "B3Generate.h"
-#include "B3InsertionSet.h"
-#include "B3SIMDValue.h"
-#include "B3StackmapGenerationParams.h"
-#include "B3SwitchValue.h"
-#include "B3UpsilonValue.h"
-#include "B3Validate.h"
-#include "B3ValueInlines.h"
-#include "B3ValueKey.h"
-#include "B3Variable.h"
-#include "B3VariableValue.h"
-#include "B3WasmAddressValue.h"
-#include "B3WasmBoundsCheckValue.h"
-#include "CompilerTimingScope.h"
-#include "FunctionAllowlist.h"
-#include "JSCJSValueInlines.h"
-#include "JSWebAssemblyArrayInlines.h"
-#include "JSWebAssemblyInstance.h"
-#include "JSWebAssemblyStruct.h"
-#include "ProbeContext.h"
-#include "ScratchRegisterAllocator.h"
-#include "WasmBaselineData.h"
-#include "WasmBranchHints.h"
-#include "WasmCallProfile.h"
-#include "WasmCallingConvention.h"
-#include "WasmContext.h"
-#include "WasmExceptionType.h"
-#include "WasmFaultSignalHandler.h"
-#include "WasmFunctionParser.h"
-#include "WasmIRGeneratorHelpers.h"
-#include "WasmMemory.h"
-#include "WasmMergedProfile.h"
-#include "WasmOSREntryData.h"
-#include "WasmOpcodeOrigin.h"
-#include "WasmOperations.h"
-#include "WasmSIMDOpcodes.h"
-#include "WasmThunks.h"
-#include "WasmTypeDefinitionInlines.h"
-#include "WebAssemblyFunctionBase.h"
-#include <limits>
-#include <wtf/FastMalloc.h>
-#include <wtf/StdLibExtras.h>
-#include <wtf/TZoneMallocInlines.h>
-#include <wtf/text/MakeString.h>
-
-#if !ENABLE(WEBASSEMBLY)
-#error ENABLE(WEBASSEMBLY_OMGJIT) is enabled, but ENABLE(WEBASSEMBLY) is not.
-#endif
-
 WTF_ALLOW_UNSAFE_BUFFER_USAGE_BEGIN
 
-#if USE(JSVALUE32_64)
-
 #define OMG_JSVALUE_32_64_PINNED_MEMORY_REGISTERS 0
-#define OMG_JSVALUE_32_64_NYI 1
-
-namespace JSC { namespace Wasm {
-
-using namespace B3;
-
-namespace {
-namespace WasmOMGIRGeneratorInternal {
-static constexpr bool verbose = false;
-static constexpr bool verboseInlining = false;
-static constexpr bool traceExecution = false;
-static constexpr bool traceStackValues = false;
-static constexpr bool verboseTailCalls = false;
-#if ASSERT_ENABLED
-static constexpr bool traceExecutionIncludesConstructionSite = false;
-#endif
-}
-}
-
-#define TRACE_VALUE(...) do { if constexpr (WasmOMGIRGeneratorInternal::traceExecution) { traceValue(__VA_ARGS__); } } while (0)
-
-#define TRACE_CF(...) do { if constexpr (WasmOMGIRGeneratorInternal::traceExecution) { traceCF(__VA_ARGS__); } } while (0)
-
-class OMGIRGenerator {
-    WTF_MAKE_TZONE_ALLOCATED(OMGIRGenerator);
-public:
-    using ExpressionType = Variable*;
-    using ResultList = Vector<ExpressionType, 8>;
-    using CallType = CallLinkInfo::CallType;
-    using CallPatchpointData = std::tuple<B3::PatchpointValue*, RefPtr<PatchpointExceptionHandle>, RefPtr<B3::StackmapGenerator>>;
-    using WasmConstRefValue = Const64Value;
-
-    static constexpr bool shouldFuseBranchCompare = false;
-    static constexpr bool tierSupportsSIMD() { return true; }
-    static constexpr bool validateFunctionBodySize = true;
-
-    struct ControlData {
-        ControlData(Procedure& proc, Origin origin, BlockSignature signature, BlockType type, unsigned stackSize, BasicBlock* continuation, BasicBlock* special = nullptr)
-            : controlBlockType(type)
-            , m_signature(signature)
-            , m_stackSize(stackSize)
-            , continuation(continuation)
-            , special(special)
-        {
-            ASSERT(type != BlockType::Try && type != BlockType::Catch);
-            if (type != BlockType::TopLevel)
-                m_stackSize -= signature.m_signature->argumentCount();
-
-            if (type == BlockType::Loop) {
-                for (unsigned i = 0; i < signature.m_signature->argumentCount(); ++i)
-                    phis.append(proc.add<Value>(Phi, toB3Type(signature.m_signature->argumentType(i)), origin));
-            } else {
-                for (unsigned i = 0; i < signature.m_signature->returnCount(); ++i)
-                    phis.append(proc.add<Value>(Phi, toB3Type(signature.m_signature->returnType(i)), origin));
-            }
-        }
-
-        ControlData(Procedure& proc, Origin origin, BlockSignature signature, BlockType type, unsigned stackSize, BasicBlock* continuation, unsigned tryStart, unsigned tryDepth)
-            : controlBlockType(type)
-            , m_signature(signature)
-            , m_stackSize(stackSize)
-            , continuation(continuation)
-            , special(nullptr)
-            , m_tryStart(tryStart)
-            , m_tryCatchDepth(tryDepth)
-        {
-            ASSERT(type == BlockType::Try || type == BlockType::TryTable);
-            m_stackSize -= signature.m_signature->argumentCount();
-            for (unsigned i = 0; i < signature.m_signature->returnCount(); ++i)
-                phis.append(proc.add<Value>(Phi, toB3Type(signature.m_signature->returnType(i)), origin));
-        }
-
-        ControlData()
-        {
-        }
-
-        static bool isIf(const ControlData& control) { return control.blockType() == BlockType::If; }
-        static bool isTry(const ControlData& control) { return control.blockType() == BlockType::Try; }
-        static bool isTryTable(const ControlData& control) { return control.blockType() == BlockType::TryTable; }
-        static bool isAnyCatch(const ControlData& control) { return control.blockType() == BlockType::Catch; }
-        static bool isTopLevel(const ControlData& control) { return control.blockType() == BlockType::TopLevel; }
-        static bool isLoop(const ControlData& control) { return control.blockType() == BlockType::Loop; }
-        static bool isBlock(const ControlData& control) { return control.blockType() == BlockType::Block; }
-        static bool isCatch(const ControlData& control)
-        {
-            if (control.blockType() != BlockType::Catch)
-                return false;
-            return control.catchKind() == CatchKind::Catch;
-        }
-
-        void dump(PrintStream& out) const
-        {
-            switch (blockType()) {
-            case BlockType::If:
-                out.print("If:       ");
-                break;
-            case BlockType::Block:
-                out.print("Block:    ");
-                break;
-            case BlockType::Loop:
-                out.print("Loop:     ");
-                break;
-            case BlockType::TopLevel:
-                out.print("TopLevel: ");
-                break;
-            case BlockType::Try:
-                out.print("Try: ");
-                break;
-            case BlockType::TryTable:
-                out.print("TryTable: ");
-                break;
-            case BlockType::Catch:
-                out.print("Catch: ");
-                break;
-            }
-            out.print("Continuation: ", *continuation, ", Special: ");
-            if (special)
-                out.print(*special);
-            else
-                out.print("None");
-        }
-
-        BlockType blockType() const { return controlBlockType; }
-
-        BlockSignature signature() const { return m_signature; }
-
-        bool hasNonVoidresult() const { return m_signature.m_signature->returnsVoid(); }
-
-        BasicBlock* targetBlockForBranch()
-        {
-            if (blockType() == BlockType::Loop)
-                return special;
-            return continuation;
-        }
-
-        void convertIfToBlock()
-        {
-            ASSERT(blockType() == BlockType::If);
-            controlBlockType = BlockType::Block;
-            special = nullptr;
-        }
-
-        void convertTryToCatch(unsigned tryEndCallSiteIndex, Variable* exception)
-        {
-            ASSERT(blockType() == BlockType::Try);
-            controlBlockType = BlockType::Catch;
-            m_catchKind = CatchKind::Catch;
-            m_tryEnd = tryEndCallSiteIndex;
-            m_exception = exception;
-        }
-
-        void convertTryToCatchAll(unsigned tryEndCallSiteIndex, Variable* exception)
-        {
-            ASSERT(blockType() == BlockType::Try);
-            controlBlockType = BlockType::Catch;
-            m_catchKind = CatchKind::CatchAll;
-            m_tryEnd = tryEndCallSiteIndex;
-            m_exception = exception;
-        }
-
-        struct TryTableTarget {
-            CatchKind type;
-            uint32_t tag;
-            const TypeDefinition* exceptionSignature;
-            ControlRef target;
-        };
-        using TargetList = Vector<TryTableTarget>;
-
-        void setTryTableTargets(TargetList&& targets)
-        {
-            m_tryTableTargets = WTFMove(targets);
-        }
-
-        void endTryTable(unsigned tryEndCallSiteIndex)
-        {
-            ASSERT(blockType() == BlockType::TryTable);
-            m_tryEnd = tryEndCallSiteIndex;
-        }
-
-        FunctionArgCount branchTargetArity() const
-        {
-            if (blockType() == BlockType::Loop)
-                return m_signature.m_signature->argumentCount();
-            return m_signature.m_signature->returnCount();
-        }
-
-        Type branchTargetType(unsigned i) const
-        {
-            ASSERT(i < branchTargetArity());
-            if (blockType() == BlockType::Loop)
-                return m_signature.m_signature->argumentType(i);
-            return m_signature.m_signature->returnType(i);
-        }
-
-        unsigned tryStart() const
-        {
-            ASSERT(controlBlockType == BlockType::Try || controlBlockType == BlockType::TryTable || controlBlockType == BlockType::Catch);
-            return m_tryStart;
-        }
-
-        unsigned tryEnd() const
-        {
-            ASSERT(controlBlockType == BlockType::Catch || controlBlockType == BlockType::TryTable);
-            return m_tryEnd;
-        }
-
-        unsigned tryDepth() const
-        {
-            ASSERT(controlBlockType == BlockType::Try || controlBlockType == BlockType::TryTable || controlBlockType == BlockType::Catch);
-            return m_tryCatchDepth;
-        }
-
-        CatchKind catchKind() const
-        {
-            ASSERT(controlBlockType == BlockType::Catch);
-            return m_catchKind;
-        }
-
-        Variable* exception() const
-        {
-            ASSERT(controlBlockType == BlockType::Catch || controlBlockType == BlockType::TryTable);
-            return m_exception;
-        }
-
-        unsigned stackSize() const { return m_stackSize; }
-
-    private:
-        // FIXME: Compress OMGIRGenerator::ControlData fields using an union
-        // https://bugs.webkit.org/show_bug.cgi?id=231212
-        friend class OMGIRGenerator;
-        BlockType controlBlockType;
-        BlockSignature m_signature;
-        unsigned m_stackSize;
-        BasicBlock* continuation;
-        BasicBlock* special;
-        Vector<Value*> phis;
-        unsigned m_tryStart;
-        unsigned m_tryEnd;
-        unsigned m_tryCatchDepth;
-        CatchKind m_catchKind;
-        Variable* m_exception;
-        TargetList m_tryTableTargets;
-    };
-
-    using ControlType = ControlData;
-    using ExpressionList = Vector<ExpressionType, 1>;
-
-    using ControlEntry = FunctionParser<OMGIRGenerator>::ControlEntry;
-    using ControlStack = FunctionParser<OMGIRGenerator>::ControlStack;
-    using Stack = FunctionParser<OMGIRGenerator>::Stack;
-    using TypedExpression = FunctionParser<OMGIRGenerator>::TypedExpression;
-    using CatchHandler = FunctionParser<OMGIRGenerator>::CatchHandler;
-    using ArgumentList = FunctionParser<OMGIRGenerator>::ArgumentList;
-
-    static_assert(std::is_same_v<ResultList, FunctionParser<OMGIRGenerator>::ResultList>);
-
-    typedef String ErrorType;
-    typedef Unexpected<ErrorType> UnexpectedResult;
-    typedef Expected<std::unique_ptr<InternalFunction>, ErrorType> Result;
-    typedef Expected<void, ErrorType> PartialResult;
-
-    static ExpressionType emptyExpression() { return nullptr; };
-
-    enum class CastKind { Cast, Test };
-
-    template <typename ...Args>
-    NEVER_INLINE UnexpectedResult WARN_UNUSED_RETURN fail(Args... args) const
-    {
-        using namespace FailureHelper; // See ADL comment in WasmParser.h.
-        return UnexpectedResult(makeString("WebAssembly.Module failed compiling: "_s, makeString(args)...));
-    }
-#define WASM_COMPILE_FAIL_IF(condition, ...) do { \
-        if (condition) [[unlikely]]                  \
-            return fail(__VA_ARGS__);             \
-    } while (0)
-
-    unsigned advanceCallSiteIndex()
-    {
-        if (m_inlineParent)
-            return m_inlineRoot->advanceCallSiteIndex();
-        return ++m_callSiteIndex;
-    }
-
-    unsigned callSiteIndex() const
-    {
-        if (m_inlineParent)
-            return m_inlineRoot->callSiteIndex();
-        return m_callSiteIndex;
-    }
-
-    OMGIRGenerator(AbstractHeapRepository&, CompilationContext&, Module&, CalleeGroup&, const ModuleInformation&, IPIntCallee&, OptimizingJITCallee&, Procedure&, Vector<UnlinkedWasmToWasmCall>&, FixedBitVector& outgoingDirectCallees, unsigned& osrEntryScratchBufferSize, MemoryMode, CompilationMode, unsigned functionIndex, unsigned loopIndexForOSREntry);
-    OMGIRGenerator(AbstractHeapRepository&, CompilationContext&, OMGIRGenerator& inlineCaller, OMGIRGenerator& inlineRoot, Module&, CalleeGroup&, unsigned functionIndex, IPIntCallee&, BasicBlock* returnContinuation, Vector<Value*> args);
-
-    void computeStackCheckSize(bool& needsOverflowCheck, int32_t& checkSize);
-
-    Value* truncate(Value *v)
-    {
-        return m_currentBlock->appendNew<Value>(m_proc, Trunc, origin(), v);
-    }
-
-    Value* wasmRefOfCell(Value* cell)
-    {
-        return m_currentBlock->appendNew<Value>(m_proc, Stitch, origin(), cell, constant(Int32, JSValue::CellTag));
-    }
-
-    Value* pointerOfWasmRef(Value* ref)
-    {
-        return m_currentBlock->appendNew<Value>(m_proc, Trunc, origin(), ref);
-    }
-
-    Value* pointerOfInt32(Value* value)
-    {
-        return value;
-    }
-
-    Value* int32OfPointer(Value* value)
-    {
-        return value;
-    }
-
-    // SIMD
-    bool usesSIMD() { return m_info.usesSIMD(m_functionIndex); }
-    void notifyFunctionUsesSIMD() { ASSERT(m_info.usesSIMD(m_functionIndex)); }
-    PartialResult WARN_UNUSED_RETURN addSIMDLoad(ExpressionType pointer, uint32_t offset, ExpressionType& result);
-    PartialResult WARN_UNUSED_RETURN addSIMDStore(ExpressionType value, ExpressionType pointer, uint32_t offset);
-    PartialResult WARN_UNUSED_RETURN addSIMDSplat(SIMDLane, ExpressionType scalar, ExpressionType& result);
-    PartialResult WARN_UNUSED_RETURN addSIMDShuffle(v128_t imm, ExpressionType a, ExpressionType b, ExpressionType& result);
-    PartialResult WARN_UNUSED_RETURN addSIMDShift(SIMDLaneOperation, SIMDInfo, ExpressionType v, ExpressionType shift, ExpressionType& result);
-    PartialResult WARN_UNUSED_RETURN addSIMDExtmul(SIMDLaneOperation, SIMDInfo, ExpressionType lhs, ExpressionType rhs, ExpressionType& result);
-    PartialResult WARN_UNUSED_RETURN addSIMDLoadSplat(SIMDLaneOperation, ExpressionType pointer, uint32_t offset, ExpressionType& result);
-    PartialResult WARN_UNUSED_RETURN addSIMDLoadLane(SIMDLaneOperation, ExpressionType pointer, ExpressionType vector, uint32_t offset, uint8_t laneIndex, ExpressionType& result);
-    PartialResult WARN_UNUSED_RETURN addSIMDStoreLane(SIMDLaneOperation, ExpressionType pointer, ExpressionType vector, uint32_t offset, uint8_t laneIndex);
-    PartialResult WARN_UNUSED_RETURN addSIMDLoadExtend(SIMDLaneOperation, ExpressionType pointer, uint32_t offset, ExpressionType& result);
-    PartialResult WARN_UNUSED_RETURN addSIMDLoadPad(SIMDLaneOperation, ExpressionType pointer, uint32_t offset, ExpressionType& result);
-
-    ExpressionType WARN_UNUSED_RETURN addConstant(v128_t value)
-    {
-        return push(m_currentBlock->appendNew<Const128Value>(m_proc, origin(), value));
-    }
-
-    // SIMD generated
-
-    #define B3_OP_CASE(OP) \
-    else if (op == SIMDLaneOperation::OP) b3Op = B3::Vector##OP;
-
-    #define B3_OP_CASES() \
-    B3::Opcode b3Op = B3::Oops; \
-    if (false) { }
-
-    auto addExtractLane(SIMDInfo info, uint8_t lane, ExpressionType v, ExpressionType& result) -> PartialResult
-    {
-        result = push(m_currentBlock->appendNew<SIMDValue>(m_proc, origin(), B3::VectorExtractLane, toB3Type(simdScalarType(info.lane)), info,
-            lane,
-            get(v)));
-        return { };
-    }
-
-    auto addReplaceLane(SIMDInfo info, uint8_t lane, ExpressionType v, ExpressionType s, ExpressionType& result) -> PartialResult
-    {
-        result = push(m_currentBlock->appendNew<SIMDValue>(m_proc, origin(), B3::VectorReplaceLane, B3::V128, info,
-            lane,
-            get(v),
-            get(s)));
-        return { };
-    }
-
-    auto addSIMDI_V(SIMDLaneOperation op, SIMDInfo info, ExpressionType v, ExpressionType& result) -> PartialResult
-    {
-        B3_OP_CASES()
-        B3_OP_CASE(Bitmask)
-        B3_OP_CASE(AnyTrue)
-        B3_OP_CASE(AllTrue)
-        result = push(m_currentBlock->appendNew<SIMDValue>(m_proc, origin(), b3Op, B3::Int32, info,
-            get(v)));
-        return { };
-    }
-
-    auto addSIMDV_V(SIMDLaneOperation op, SIMDInfo info, ExpressionType v, ExpressionType& result) -> PartialResult
-    {
-        B3_OP_CASES()
-        B3_OP_CASE(Demote)
-        B3_OP_CASE(Promote)
-        B3_OP_CASE(Abs)
-        B3_OP_CASE(Popcnt)
-        B3_OP_CASE(Ceil)
-        B3_OP_CASE(Floor)
-        B3_OP_CASE(Trunc)
-        B3_OP_CASE(Nearest)
-        B3_OP_CASE(Sqrt)
-        B3_OP_CASE(ExtaddPairwise)
-        B3_OP_CASE(Convert)
-        B3_OP_CASE(ConvertLow)
-        B3_OP_CASE(ExtendHigh)
-        B3_OP_CASE(ExtendLow)
-        B3_OP_CASE(TruncSat)
-        B3_OP_CASE(RelaxedTruncSat)
-        B3_OP_CASE(Not)
-        B3_OP_CASE(Neg)
-
-        result = push(m_currentBlock->appendNew<SIMDValue>(m_proc, origin(), b3Op, B3::V128, info,
-            get(v)));
-        return { };
-    }
-
-    auto addSIMDBitwiseSelect(ExpressionType v1, ExpressionType v2, ExpressionType c, ExpressionType& result) -> PartialResult
-    {
-        auto b3Op = B3::VectorBitwiseSelect;
-        result = push(m_currentBlock->appendNew<SIMDValue>(m_proc, origin(), b3Op, B3::V128, SIMDInfo { SIMDLane::v128, SIMDSignMode::None },
-            get(v1), get(v2), get(c)));
-        return { };
-    }
-
-    auto addSIMDRelOp(SIMDLaneOperation, SIMDInfo info, ExpressionType lhs, ExpressionType rhs, Air::Arg relOp, ExpressionType& result) -> PartialResult
-    {
-        B3::Opcode b3Op = Oops;
-        if (scalarTypeIsIntegral(info.lane)) {
-            switch (relOp.asRelationalCondition()) {
-            case MacroAssembler::Equal:
-                b3Op = VectorEqual;
-                break;
-            case MacroAssembler::NotEqual:
-                b3Op = VectorNotEqual;
-                break;
-            case MacroAssembler::LessThan:
-                b3Op = VectorLessThan;
-                break;
-            case MacroAssembler::LessThanOrEqual:
-                b3Op = VectorLessThanOrEqual;
-                break;
-            case MacroAssembler::Below:
-                b3Op = VectorBelow;
-                break;
-            case MacroAssembler::BelowOrEqual:
-                b3Op = VectorBelowOrEqual;
-                break;
-            case MacroAssembler::GreaterThan:
-                b3Op = VectorGreaterThan;
-                break;
-            case MacroAssembler::GreaterThanOrEqual:
-                b3Op = VectorGreaterThanOrEqual;
-                break;
-            case MacroAssembler::Above:
-                b3Op = VectorAbove;
-                break;
-            case MacroAssembler::AboveOrEqual:
-                b3Op = VectorAboveOrEqual;
-                break;
-            }
-        } else {
-            switch (relOp.asDoubleCondition()) {
-            case MacroAssembler::DoubleEqualAndOrdered:
-                b3Op = VectorEqual;
-                break;
-            case MacroAssembler::DoubleNotEqualOrUnordered:
-                b3Op = VectorNotEqual;
-                break;
-            case MacroAssembler::DoubleLessThanAndOrdered:
-                b3Op = VectorLessThan;
-                break;
-            case MacroAssembler::DoubleLessThanOrEqualAndOrdered:
-                b3Op = VectorLessThanOrEqual;
-                break;
-            case MacroAssembler::DoubleGreaterThanAndOrdered:
-                b3Op = VectorGreaterThan;
-                break;
-            case MacroAssembler::DoubleGreaterThanOrEqualAndOrdered:
-                b3Op = VectorGreaterThanOrEqual;
-                break;
-            default:
-                RELEASE_ASSERT_NOT_REACHED();
-                break;
-            }
-        }
-
-        result = push(m_currentBlock->appendNew<SIMDValue>(m_proc, origin(), b3Op, B3::V128, info,
-            get(lhs), get(rhs)));
-        return { };
-    }
-
-    Value* fixupOutOfBoundsIndicesForSwizzle(Value* input, Value* indexes)
-    {
-        // The intel version of the swizzle instruction does not handle OOB indices properly,
-        // so we need to fix them up.
-        ASSERT(isX86());
-        // Let each byte mask be 112 (0x70) then after VectorAddSat
-        // each index > 15 would set the saturated index's bit 7 to 1,
-        // whose corresponding byte will be zero cleared in VectorSwizzle.
-        // https://github.com/WebAssembly/simd/issues/93
-        v128_t mask;
-        mask.u64x2[0] = 0x7070707070707070;
-        mask.u64x2[1] = 0x7070707070707070;
-        auto saturatingMask = m_currentBlock->appendNew<Const128Value>(m_proc, origin(), mask);
-        auto saturatedIndexes = m_currentBlock->appendNew<SIMDValue>(m_proc, origin(), VectorAddSat, B3::V128, SIMDLane::i8x16, SIMDSignMode::Unsigned, saturatingMask, indexes);
-        return m_currentBlock->appendNew<SIMDValue>(m_proc, origin(), VectorSwizzle, B3::V128, SIMDLane::i8x16, SIMDSignMode::None, input, saturatedIndexes);
-    }
-
-    auto addSIMDV_VV(SIMDLaneOperation op, SIMDInfo info, ExpressionType a, ExpressionType b, ExpressionType& result) -> PartialResult
-    {
-        B3_OP_CASES()
-        B3_OP_CASE(And)
-        B3_OP_CASE(Andnot)
-        B3_OP_CASE(AvgRound)
-        B3_OP_CASE(DotProduct)
-        B3_OP_CASE(Add)
-        B3_OP_CASE(Mul)
-        B3_OP_CASE(MulSat)
-        B3_OP_CASE(Sub)
-        B3_OP_CASE(Div)
-        B3_OP_CASE(Pmax)
-        B3_OP_CASE(Pmin)
-        B3_OP_CASE(Or)
-        B3_OP_CASE(Swizzle)
-        B3_OP_CASE(RelaxedSwizzle)
-        B3_OP_CASE(Xor)
-        B3_OP_CASE(Narrow)
-        B3_OP_CASE(AddSat)
-        B3_OP_CASE(SubSat)
-        B3_OP_CASE(Max)
-        B3_OP_CASE(Min)
-
-        if (isX86() && b3Op == B3::VectorSwizzle) {
-            result = push(fixupOutOfBoundsIndicesForSwizzle(get(a), get(b)));
-            return { };
-        }
-
-        result = push(m_currentBlock->appendNew<SIMDValue>(m_proc, origin(), b3Op, B3::V128, info,
-            get(a), get(b)));
-        return { };
-    }
-
-    auto addSIMDRelaxedFMA(SIMDLaneOperation op, SIMDInfo info, ExpressionType m1, ExpressionType m2, ExpressionType add, ExpressionType& result) -> PartialResult
-    {
-        B3_OP_CASES()
-        B3_OP_CASE(RelaxedMAdd)
-        B3_OP_CASE(RelaxedNMAdd)
-
-        result = push(m_currentBlock->appendNew<SIMDValue>(m_proc, origin(), b3Op, B3::V128, info,
-            get(m1), get(m2), get(add)));
-        return { };
-    }
-
-    PartialResult WARN_UNUSED_RETURN addDrop(ExpressionType);
-    PartialResult WARN_UNUSED_RETURN addInlinedArguments(const TypeDefinition&);
-    PartialResult WARN_UNUSED_RETURN addArguments(const TypeDefinition&);
-    PartialResult WARN_UNUSED_RETURN addLocal(Type, uint32_t);
-    ExpressionType addConstant(Type, uint64_t);
-
-    // References
-    PartialResult WARN_UNUSED_RETURN addRefIsNull(ExpressionType value, ExpressionType& result);
-    PartialResult WARN_UNUSED_RETURN addRefFunc(FunctionSpaceIndex index, ExpressionType& result);
-    PartialResult WARN_UNUSED_RETURN addRefAsNonNull(TypedExpression, ExpressionType&);
-    PartialResult WARN_UNUSED_RETURN addRefEq(ExpressionType, ExpressionType, ExpressionType&);
-
-    // Tables
-    PartialResult WARN_UNUSED_RETURN addTableGet(unsigned, ExpressionType index, ExpressionType& result);
-    PartialResult WARN_UNUSED_RETURN addTableSet(unsigned, ExpressionType index, ExpressionType value);
-    PartialResult WARN_UNUSED_RETURN addTableInit(unsigned, unsigned, ExpressionType dstOffset, ExpressionType srcOffset, ExpressionType length);
-    PartialResult WARN_UNUSED_RETURN addElemDrop(unsigned);
-    PartialResult WARN_UNUSED_RETURN addTableSize(unsigned, ExpressionType& result);
-    PartialResult WARN_UNUSED_RETURN addTableGrow(unsigned, ExpressionType fill, ExpressionType delta, ExpressionType& result);
-    PartialResult WARN_UNUSED_RETURN addTableFill(unsigned, ExpressionType offset, ExpressionType fill, ExpressionType count);
-    PartialResult WARN_UNUSED_RETURN addTableCopy(unsigned, unsigned, ExpressionType dstOffset, ExpressionType srcOffset, ExpressionType length);
-
-    // Locals
-    PartialResult WARN_UNUSED_RETURN getLocal(uint32_t index, ExpressionType& result);
-    PartialResult WARN_UNUSED_RETURN setLocal(uint32_t index, ExpressionType value);
-    PartialResult WARN_UNUSED_RETURN teeLocal(uint32_t, ExpressionType, ExpressionType& result);
-
-    // Globals
-    PartialResult WARN_UNUSED_RETURN getGlobal(uint32_t index, ExpressionType& result);
-    PartialResult WARN_UNUSED_RETURN setGlobal(uint32_t index, ExpressionType value);
-
-    // Memory
-    PartialResult WARN_UNUSED_RETURN load(LoadOpType, ExpressionType pointer, ExpressionType& result, uint32_t offset);
-    PartialResult WARN_UNUSED_RETURN store(StoreOpType, ExpressionType pointer, ExpressionType value, uint32_t offset);
-    PartialResult WARN_UNUSED_RETURN addGrowMemory(ExpressionType delta, ExpressionType& result);
-    PartialResult WARN_UNUSED_RETURN addCurrentMemory(ExpressionType& result);
-    PartialResult WARN_UNUSED_RETURN addMemoryFill(ExpressionType dstAddress, ExpressionType targetValue, ExpressionType count);
-    PartialResult WARN_UNUSED_RETURN addMemoryCopy(ExpressionType dstAddress, ExpressionType srcAddress, ExpressionType count);
-    PartialResult WARN_UNUSED_RETURN addMemoryInit(unsigned, ExpressionType dstAddress, ExpressionType srcAddress, ExpressionType length);
-    PartialResult WARN_UNUSED_RETURN addDataDrop(unsigned);
-
-    // Atomics
-    PartialResult WARN_UNUSED_RETURN atomicLoad(ExtAtomicOpType, Type, ExpressionType pointer, ExpressionType& result, uint32_t offset);
-    PartialResult WARN_UNUSED_RETURN atomicStore(ExtAtomicOpType, Type, ExpressionType pointer, ExpressionType value, uint32_t offset);
-    PartialResult WARN_UNUSED_RETURN atomicBinaryRMW(ExtAtomicOpType, Type, ExpressionType pointer, ExpressionType value, ExpressionType& result, uint32_t offset);
-    PartialResult WARN_UNUSED_RETURN atomicCompareExchange(ExtAtomicOpType, Type, ExpressionType pointer, ExpressionType expected, ExpressionType value, ExpressionType& result, uint32_t offset);
-
-    PartialResult WARN_UNUSED_RETURN atomicWait(ExtAtomicOpType, ExpressionType pointer, ExpressionType value, ExpressionType timeout, ExpressionType& result, uint32_t offset);
-    PartialResult WARN_UNUSED_RETURN atomicNotify(ExtAtomicOpType, ExpressionType pointer, ExpressionType value, ExpressionType& result, uint32_t offset);
-    PartialResult WARN_UNUSED_RETURN atomicFence(ExtAtomicOpType, uint8_t flags);
-
-    // Saturated truncation.
-    PartialResult WARN_UNUSED_RETURN truncSaturated(Ext1OpType, ExpressionType operand, ExpressionType& result, Type returnType, Type operandType);
-
-    // GC
-    PartialResult WARN_UNUSED_RETURN addRefI31(ExpressionType value, ExpressionType& result);
-    PartialResult WARN_UNUSED_RETURN addI31GetS(TypedExpression ref, ExpressionType& result);
-    PartialResult WARN_UNUSED_RETURN addI31GetU(TypedExpression ref, ExpressionType& result);
-    PartialResult WARN_UNUSED_RETURN addArrayNew(uint32_t index, ExpressionType size, ExpressionType value, ExpressionType& result);
-    PartialResult WARN_UNUSED_RETURN addArrayNewDefault(uint32_t index, ExpressionType size, ExpressionType& result);
-    PartialResult WARN_UNUSED_RETURN addArrayNewFixed(uint32_t typeIndex, ArgumentList& args, ExpressionType& result);
-    PartialResult WARN_UNUSED_RETURN addArrayGet(ExtGCOpType arrayGetKind, uint32_t typeIndex, TypedExpression arrayref, ExpressionType index, ExpressionType& result);
-    PartialResult WARN_UNUSED_RETURN addArrayNewData(uint32_t typeIndex, uint32_t dataIndex, ExpressionType size, ExpressionType offset, ExpressionType& result);
-    PartialResult WARN_UNUSED_RETURN addArrayNewElem(uint32_t typeIndex, uint32_t elemSegmentIndex, ExpressionType size, ExpressionType offset, ExpressionType& result);
-    PartialResult WARN_UNUSED_RETURN addArraySet(uint32_t typeIndex, TypedExpression arrayref, ExpressionType index, ExpressionType value);
-    PartialResult WARN_UNUSED_RETURN addArrayLen(TypedExpression arrayref, ExpressionType& result);
-    PartialResult WARN_UNUSED_RETURN addArrayFill(uint32_t, TypedExpression, ExpressionType, ExpressionType, ExpressionType);
-    PartialResult WARN_UNUSED_RETURN addArrayCopy(uint32_t, TypedExpression, ExpressionType, uint32_t, TypedExpression, ExpressionType, ExpressionType);
-    PartialResult WARN_UNUSED_RETURN addArrayInitElem(uint32_t, TypedExpression, ExpressionType, uint32_t, ExpressionType, ExpressionType);
-    PartialResult WARN_UNUSED_RETURN addArrayInitData(uint32_t, TypedExpression, ExpressionType, uint32_t, ExpressionType, ExpressionType);
-    PartialResult WARN_UNUSED_RETURN addStructNew(uint32_t typeIndex, ArgumentList& args, ExpressionType& result);
-    PartialResult WARN_UNUSED_RETURN addStructNewDefault(uint32_t index, ExpressionType& result);
-    PartialResult WARN_UNUSED_RETURN addStructGet(ExtGCOpType structGetKind, TypedExpression structReference, const StructType&, uint32_t fieldIndex, ExpressionType& result);
-    PartialResult WARN_UNUSED_RETURN addStructSet(TypedExpression structReference, const StructType&, uint32_t fieldIndex, ExpressionType value);
-    PartialResult WARN_UNUSED_RETURN addRefTest(TypedExpression reference, bool allowNull, int32_t heapType, bool shouldNegate, ExpressionType& result);
-    PartialResult WARN_UNUSED_RETURN addRefCast(TypedExpression reference, bool allowNull, int32_t heapType, ExpressionType& result);
-    PartialResult WARN_UNUSED_RETURN addAnyConvertExtern(ExpressionType reference, ExpressionType& result);
-    PartialResult WARN_UNUSED_RETURN addExternConvertAny(ExpressionType reference, ExpressionType& result);
-
-    // Basic operators
-#define X(name, opcode, short, idx, ...) \
-    PartialResult WARN_UNUSED_RETURN add##name(ExpressionType arg, ExpressionType& result);
-    FOR_EACH_WASM_UNARY_OP(X)
-#undef X
-#define X(name, opcode, short, idx, ...) \
-    PartialResult WARN_UNUSED_RETURN add##name(ExpressionType left, ExpressionType right, ExpressionType& result);
-    FOR_EACH_WASM_BINARY_OP(X)
-#undef X
-
-    PartialResult WARN_UNUSED_RETURN addSelect(ExpressionType condition, ExpressionType nonZero, ExpressionType zero, ExpressionType& result);
-
-    // Control flow
-    ControlData WARN_UNUSED_RETURN addTopLevel(BlockSignature);
-    PartialResult WARN_UNUSED_RETURN addBlock(BlockSignature, Stack& enclosingStack, ControlType& newBlock, Stack& newStack);
-    PartialResult WARN_UNUSED_RETURN addLoop(BlockSignature, Stack& enclosingStack, ControlType& block, Stack& newStack, uint32_t loopIndex);
-    PartialResult WARN_UNUSED_RETURN addIf(ExpressionType condition, BlockSignature, Stack& enclosingStack, ControlType& result, Stack& newStack);
-    PartialResult WARN_UNUSED_RETURN addElse(ControlData&, const Stack&);
-    PartialResult WARN_UNUSED_RETURN addElseToUnreachable(ControlData&);
-
-    PartialResult WARN_UNUSED_RETURN addTry(BlockSignature, Stack& enclosingStack, ControlType& result, Stack& newStack);
-    PartialResult WARN_UNUSED_RETURN addTryTable(BlockSignature, Stack& enclosingStack, const Vector<CatchHandler>& targets, ControlType& result, Stack& newStack);
-    PartialResult WARN_UNUSED_RETURN addCatch(unsigned exceptionIndex, const TypeDefinition&, Stack&, ControlType&, ResultList&);
-    PartialResult WARN_UNUSED_RETURN addCatchToUnreachable(unsigned exceptionIndex, const TypeDefinition&, ControlType&, ResultList&);
-    PartialResult WARN_UNUSED_RETURN addCatchAll(Stack&, ControlType&);
-    PartialResult WARN_UNUSED_RETURN addCatchAllToUnreachable(ControlType&);
-    PartialResult WARN_UNUSED_RETURN addDelegate(ControlType&, ControlType&);
-    PartialResult WARN_UNUSED_RETURN addDelegateToUnreachable(ControlType&, ControlType&);
-    PartialResult WARN_UNUSED_RETURN addThrow(unsigned exceptionIndex, ArgumentList& args, Stack&);
-    PartialResult WARN_UNUSED_RETURN addRethrow(unsigned, ControlType&);
-    PartialResult WARN_UNUSED_RETURN addThrowRef(TypedExpression exception, Stack&);
-
-    PartialResult WARN_UNUSED_RETURN addInlinedReturn(const auto& returnValues);
-
-    PartialResult WARN_UNUSED_RETURN addReturn(const ControlData&, const Stack& returnValues);
-    PartialResult WARN_UNUSED_RETURN addBranch(ControlData&, ExpressionType condition, const Stack& returnValues);
-    PartialResult WARN_UNUSED_RETURN addBranchNull(ControlType&, ExpressionType, const Stack&, bool, ExpressionType&);
-    PartialResult WARN_UNUSED_RETURN addBranchCast(ControlType&, TypedExpression, const Stack&, bool, int32_t, bool);
-    PartialResult WARN_UNUSED_RETURN addSwitch(ExpressionType condition, const Vector<ControlData*>& targets, ControlData& defaultTargets, const Stack& expressionStack);
-    PartialResult WARN_UNUSED_RETURN endBlock(ControlEntry&, Stack& expressionStack);
-    PartialResult WARN_UNUSED_RETURN addEndToUnreachable(ControlEntry&, const Stack& = { });
-
-    PartialResult WARN_UNUSED_RETURN endTopLevel(BlockSignature, const Stack&) { return { }; }
-
-    // Fused comparison stubs (B3 will do this for us later).
-    PartialResult WARN_UNUSED_RETURN addFusedBranchCompare(OpType, ControlType&, ExpressionType, const Stack&) { RELEASE_ASSERT_NOT_REACHED(); }
-    PartialResult WARN_UNUSED_RETURN addFusedBranchCompare(OpType, ControlType&, ExpressionType, ExpressionType, const Stack&) { RELEASE_ASSERT_NOT_REACHED(); }
-    PartialResult WARN_UNUSED_RETURN addFusedIfCompare(OpType, ExpressionType, BlockSignature, Stack&, ControlType&, Stack&) { RELEASE_ASSERT_NOT_REACHED(); }
-    PartialResult WARN_UNUSED_RETURN addFusedIfCompare(OpType, ExpressionType, ExpressionType, BlockSignature, Stack&, ControlType&, Stack&) { RELEASE_ASSERT_NOT_REACHED(); }
-
-    // Calls
-    PartialResult WARN_UNUSED_RETURN addCall(unsigned, FunctionSpaceIndex functionIndexSpace, const TypeDefinition&, ArgumentList& args, ResultList& results, CallType = CallType::Call);
-    PartialResult WARN_UNUSED_RETURN addCallIndirect(unsigned, unsigned tableIndex, const TypeDefinition&, ArgumentList& args, ResultList& results, CallType = CallType::Call);
-    PartialResult WARN_UNUSED_RETURN addCallRef(unsigned, const TypeDefinition&, ArgumentList& args, ResultList& results, CallType = CallType::Call);
-    PartialResult WARN_UNUSED_RETURN addUnreachable();
-    PartialResult WARN_UNUSED_RETURN addCrash();
-
-    using ValueResults = Vector<Value*, 16>;
-    void fillCallResults(Value* callResult, const TypeDefinition& signature, ValueResults&);
-    PartialResult WARN_UNUSED_RETURN emitDirectCall(unsigned, FunctionSpaceIndex functionIndexSpace, const TypeDefinition&, ArgumentList& args, ValueResults&, CallType = CallType::Call);
-    PartialResult WARN_UNUSED_RETURN emitIndirectCall(Value* calleeInstance, Value* calleeCode, Value* boxedCalleeCallee, const TypeDefinition&, const ArgumentList& args, ValueResults&, CallType = CallType::Call);
-
-    Vector<ConstrainedValue> createCallConstrainedArgs(BasicBlock*, const CallInformation& wasmCalleeInfo, const ArgumentList&);
-    auto createCallPatchpoint(BasicBlock*, const TypeDefinition&, const CallInformation&, const ArgumentList& tmpArgs) -> CallPatchpointData;
-    auto createTailCallPatchpoint(BasicBlock*, const TypeDefinition&, const CallInformation& wasmCallerInfoAsCallee, const CallInformation& wasmCalleeInfoAsCallee, const ArgumentList& tmpArgSourceLocations, Vector<B3::ConstrainedValue> patchArgs) -> CallPatchpointData;
-
-    bool canInline(FunctionSpaceIndex functionIndexSpace, unsigned callProfileIndex) const;
-    PartialResult WARN_UNUSED_RETURN emitInlineDirectCall(FunctionCodeIndex calleeIndex, const TypeDefinition&, ArgumentList& args, ValueResults&);
-
-    void dump(const ControlStack&, const Stack* expressionStack);
-    void setParser(FunctionParser<OMGIRGenerator>* parser) { m_parser = parser; };
-    ALWAYS_INLINE void willParseOpcode() { }
-    ALWAYS_INLINE void willParseExtendedOpcode() { }
-    ALWAYS_INLINE void didParseOpcode() { }
-    void didFinishParsingLocals() { }
-    void didPopValueFromStack(ExpressionType expr, ASCIILiteral message)
-    {
-        --m_stackSize;
-        TRACE_VALUE(Wasm::Types::Void, get(expr), "pop at height: ", m_stackSize.value() + 1, " site: [", message, "], var ", *expr);
-    }
-    const Ref<TypeDefinition> getTypeDefinition(uint32_t typeIndex) { return m_info.typeSignatures[typeIndex]; }
-    const ArrayType* getArrayTypeDefinition(uint32_t);
-    void getArrayElementType(uint32_t, StorageType&);
-    void getArrayRefType(uint32_t, Type&);
-
-    Value* constant(B3::Type, uint64_t bits, std::optional<Origin> = std::nullopt);
-    Value* constant(B3::Type, v128_t bits, std::optional<Origin> = std::nullopt);
-    Value* framePointer();
-    void insertEntrySwitch();
-    void insertConstants();
-
-    B3::Type toB3ResultType(const TypeDefinition*);
-
-    void addStackMap(unsigned callSiteIndex, StackMap&& stackmap)
-    {
-        if (m_inlineParent) {
-            m_inlineRoot->addStackMap(callSiteIndex, WTFMove(stackmap));
-            return;
-        }
-        m_stackmaps.add(CallSiteIndex(callSiteIndex), WTFMove(stackmap));
-    }
-
-    StackMaps&& takeStackmaps()
-    {
-        RELEASE_ASSERT(m_inlineRoot == this);
-        return WTFMove(m_stackmaps);
-    }
-
-    Vector<UnlinkedHandlerInfo>&& takeExceptionHandlers()
-    {
-        RELEASE_ASSERT(m_inlineRoot == this);
-        return WTFMove(m_exceptionHandlers);
-    }
-
-private:
-    void emitPrepareWasmOperation(BasicBlock* block)
-    {
-#if !USE(BUILTIN_FRAME_ADDRESS) || ASSERT_ENABLED
-        // Prepare wasm operation calls.
-        block->appendNew<B3::MemoryValue>(m_proc, B3::Store, origin(), framePointer(), instanceValue(), JSWebAssemblyInstance::offsetOfTemporaryCallFrame());
-#else
-        UNUSED_PARAM(block);
-#endif
-    }
-
-    template<typename OperationType, typename ...Args>
-    Value* callWasmOperation(BasicBlock* block, B3::Type resultType, OperationType operation, Args&&... args)
-    {
-        emitPrepareWasmOperation(block);
-        Value* operationValue = block->appendNew<ConstPtrValue>(m_proc, origin(), tagCFunction<OperationPtrTag>(operation));
-        return block->appendNew<CCallValue>(m_proc, resultType, origin(), operationValue, std::forward<Args>(args)...);
-    }
-
-    void emitExceptionCheck(CCallHelpers&, Origin, ExceptionType);
-
-    void emitWriteBarrierForJSWrapper();
-    void emitWriteBarrier(Value* cell, Value* instanceCell);
-    Value* emitCheckAndPreparePointer(Value* pointer, uint32_t offset, uint32_t sizeOfOp);
-    B3::Kind memoryKind(B3::Opcode memoryOp);
-    Value* emitLoadOp(LoadOpType, Value* pointer, uint32_t offset);
-    void emitStoreOp(StoreOpType, Value* pointer, Value*, uint32_t offset);
-
-    Value* sanitizeAtomicResult(ExtAtomicOpType, Type, Value* result);
-    Value* emitAtomicLoadOp(ExtAtomicOpType, Type, Value* pointer, uint32_t offset);
-    void emitAtomicStoreOp(ExtAtomicOpType, Type, Value* pointer, Value*, uint32_t offset);
-    Value* emitAtomicBinaryRMWOp(ExtAtomicOpType, Type, Value* pointer, Value*, uint32_t offset);
-    Value* emitAtomicCompareExchange(ExtAtomicOpType, Type, Value* pointer, Value* expected, Value*, uint32_t offset);
-
-    Value* decodeNonNullStructure(Value* structureID);
-    Value* encodeStructureID(Value* structure);
-
-    Value* allocatorForWasmGCHeapCellSize(Value* size, BasicBlock* slowPath);
-    Value* allocateWasmGCHeapCell(Value* allocator, BasicBlock* slowPath);
-    Value* allocateWasmGCObject(Value* allocator, Value* structureID, Value* typeInfo, BasicBlock* slowPath);
-    Value* allocateWasmGCArrayUninitialized(uint32_t typeIndex, Value* size);
-    Value* allocateWasmGCStructUninitialized(uint32_t typeIndex);
-
-    void mutatorFence();
-
-    Value* emitGetArrayPayloadBase(Wasm::StorageType, Value*);
-    Value* emitGetArraySizeWithNullCheck(Type arrayType, Value*);
-    void emitNullCheck(Value*, ExceptionType);
-    bool emitNullCheckBeforeAccess(Value*, ptrdiff_t offset);
-    void emitArraySetUnchecked(uint32_t, Value*, Value*, Value*);
-    bool emitArraySetUncheckedWithoutWriteBarrier(uint32_t, Value*, Value*, Value*);
-    // Returns true if a writeBarrier/mutatorFence is needed.
-    bool WARN_UNUSED_RETURN emitStructSet(bool canTrap, Value*, uint32_t, const StructType&, Value*);
-    Value* WARN_UNUSED_RETURN allocateWasmGCArray(uint32_t typeIndex, Value* initValue, Value* size);
-    using ArraySegmentOperation = EncodedJSValue SYSV_ABI (&)(JSC::JSWebAssemblyInstance*, uint32_t, uint32_t, uint32_t, uint32_t);
-    ExpressionType WARN_UNUSED_RETURN pushArrayNewFromSegment(ArraySegmentOperation, uint32_t typeIndex, uint32_t segmentIndex, ExpressionType arraySize, ExpressionType offset, ExceptionType);
-    void emitRefTestOrCast(CastKind, TypedExpression, bool, int32_t, bool, ExpressionType&);
-    template <typename Generator>
-    void emitCheckOrBranchForCast(CastKind, Value*, const Generator&, BasicBlock*);
-    Value* emitLoadRTTFromObject(Value*);
-
-    void unify(Value* phi, const ExpressionType source);
-    void unifyValuesWithBlock(const Stack& resultStack, const ControlData& block);
-
-    void emitChecksForModOrDiv(B3::Opcode, Value* left, Value* right);
-
-    int32_t WARN_UNUSED_RETURN fixupPointerPlusOffset(Value*&, uint32_t);
-    Value* WARN_UNUSED_RETURN fixupPointerPlusOffsetForAtomicOps(ExtAtomicOpType, Value*, uint32_t);
-
-    void restoreWasmContextInstance(BasicBlock*, Value*);
-    void restoreWebAssemblyGlobalState(const MemoryInformation&, Value* instance, BasicBlock*);
-    void reloadMemoryRegistersFromInstance(const MemoryInformation&, Value* instance, BasicBlock*);
-
-    enum OSRBufferMode {
-        LoadI64, // Load I64 values directly, and let later passes lower them; Needed for interacting with lower tiers.
-        SplitI64, // Split I64 values into two I32 values.
-    };
-
-    Value* loadFromScratchBuffer(OSRBufferMode, unsigned& indexInBuffer, Value* pointer, B3::Type);
-    void connectControlAtEntrypoint(OSRBufferMode, unsigned& indexInBuffer, Value* pointer, ControlData&, Stack& expressionStack, ControlData& currentData, bool fillLoopPhis = false);
-    Value* emitCatchImpl(CatchKind, ControlType&, unsigned exceptionIndex = 0);
-    void emitCatchTableImpl(ControlData& entryData, const ControlData::TryTableTarget&);
-    RefPtr<PatchpointExceptionHandle> preparePatchpointForExceptions(BasicBlock*, PatchpointValue*);
-    void connectValuesForCatchEntrypoint(ControlData& catchData, Value* pointer);
-
-    Origin origin();
-
-    ExpressionType getPushVariable(B3::Type type)
-    {
-        ++m_stackSize;
-        if (m_stackSize > m_maxStackSize) {
-            m_maxStackSize = m_stackSize;
-            Variable* var = m_proc.addVariable(type);
-            if constexpr (WasmOMGIRGeneratorInternal::traceStackValues)
-                set(var, constant(type, 0xBADBEEFEF));
-            m_stack.append(var);
-            return var;
-        }
-
-        if constexpr (WasmOMGIRGeneratorInternal::traceStackValues) {
-            // When we push, everything else *should* be dead
-            for (unsigned i = m_stackSize - 1; i < m_stack.size(); ++i)
-                set(m_stack[i], constant(m_stack[i]->type(), 0xBADBEEFEF));
-        }
-
-        Variable* var = m_stack[m_stackSize - 1];
-        if (var->type() == type)
-            return var;
-
-        var = m_proc.addVariable(type);
-        m_stack[m_stackSize - 1] = var;
-        return var;
-    }
-
-    ExpressionType push(Value* value)
-    {
-        Variable* var = getPushVariable(value->type());
-        set(var, value);
-        if constexpr (!WasmOMGIRGeneratorInternal::traceExecution)
-            return var;
-        String site;
-#if ASSERT_ENABLED
-        if constexpr (WasmOMGIRGeneratorInternal::traceExecutionIncludesConstructionSite)
-            site = Value::generateCompilerConstructionSite();
-#endif
-        TRACE_VALUE(Wasm::Types::Void, get(var), "push to stack height ", m_stackSize.value(), " site: [", site, "] var ", *var);
-        return var;
-    }
-
-    Value* get(BasicBlock* block, Variable* variable)
-    {
-        return block->appendNew<VariableValue>(m_proc, B3::Get, origin(), variable);
-    }
-
-    Value* get(Variable* variable)
-    {
-        return get(m_currentBlock, variable);
-    }
-
-    Value* set(BasicBlock* block, Variable* dst, Value* src)
-    {
-        return block->appendNew<VariableValue>(m_proc, B3::Set, origin(), dst, src);
-    }
-
-    Value* set(Variable* dst, Value* src)
-    {
-        return set(m_currentBlock, dst, src);
-    }
-
-    Value* set(Variable* dst, Variable* src)
-    {
-        return set(dst, get(src));
-    }
-
-    bool useSignalingMemory() const
-    {
-        return m_mode == MemoryMode::Signaling;
-    }
-
-    template<typename... Args>
-    void traceValue(Type, Value*, Args&&... info);
-    template<typename... Args>
-    void traceCF(Args&&... info);
-
-    FunctionParser<OMGIRGenerator>* m_parser { nullptr };
-    AbstractHeapRepository& m_heaps;
-    CompilationContext& m_context;
-    Module& m_module;
-    CalleeGroup& m_calleeGroup;
-    const ModuleInformation& m_info;
-    const IPIntCallee& m_profiledCallee;
-    OptimizingJITCallee* m_callee;
-    const MemoryMode m_mode { MemoryMode::BoundsChecking };
-    const CompilationMode m_compilationMode;
-    const FunctionCodeIndex m_functionIndex;
-    const unsigned m_loopIndexForOSREntry { UINT_MAX };
-    std::unique_ptr<MergedProfile> m_profile;
-
-    struct RootBlock {
-        BasicBlock* block;
-        bool usesSIMD;
-    };
-
-    Procedure& m_proc;
-    Vector<RootBlock> m_rootBlocks;
-    BasicBlock* m_topLevelBlock;
-    BasicBlock* m_currentBlock { nullptr };
-
-    // Only used when this is an inlined context
-    BasicBlock* m_returnContinuation { nullptr };
-    OMGIRGenerator* m_inlineRoot { nullptr };
-    OMGIRGenerator* m_inlineParent { nullptr };
-    Vector<Value*> m_inlinedArgs;
-    Vector<Variable*> m_inlinedResults;
-    unsigned m_inlineDepth { 0 };
-    Checked<uint32_t> m_inlinedBytes { 0 };
-
-    Vector<Variable*> m_locals;
-    Vector<Variable*> m_stack;
-    Vector<UnlinkedWasmToWasmCall>& m_unlinkedWasmToWasmCalls; // List each call site and the function index whose address it should be patched with.
-    FixedBitVector& m_directCallees; // Note this includes call targets from functions we inline.
-    unsigned* m_osrEntryScratchBufferSize;
-    UncheckedKeyHashMap<ValueKey, Value*> m_constantPool;
-    UncheckedKeyHashMap<const TypeDefinition*, B3::Type> m_tupleMap;
-    InsertionSet m_constantInsertionValues;
-    Value* m_framePointer { nullptr };
-    bool m_makesCalls { false };
-    bool m_makesTailCalls { false };
-
-    // This tracks the maximum stack offset for a tail call, to be used in the stack overflow check.
-    Checked<int32_t> m_tailCallStackOffsetFromFP { 0 };
-
-    bool m_hasExceptionHandlers;
-
-    Value* m_instanceValue { nullptr };
-    Value* m_baseMemoryValue { nullptr };
-    Value* m_boundsCheckingSizeValue { nullptr };
-
-    Value* instanceValue()
-    {
-        return m_instanceValue;
-    }
-
-    Value* baseMemoryValue()
-    {
-        return m_baseMemoryValue;
-    }
-
-    Value* boundsCheckingSizeValue()
-    {
-        return m_boundsCheckingSizeValue;
-    }
-
-    uint32_t m_maxNumJSCallArguments { 0 };
-    unsigned m_numImportFunctions;
-
-    Checked<unsigned> m_tryCatchDepth { 0 };
-    Checked<unsigned> m_callSiteIndex { 0 };
-    Checked<unsigned> m_stackSize { 0 };
-    Checked<unsigned> m_maxStackSize { 0 };
-    StackMaps m_stackmaps;
-    Vector<UnlinkedHandlerInfo> m_exceptionHandlers;
-
-    RefPtr<B3::Air::PrologueGenerator> m_prologueGenerator;
-
-    Vector<std::unique_ptr<OMGIRGenerator>> m_protectedInlineeGenerators;
-    Vector<std::unique_ptr<FunctionParser<OMGIRGenerator>>> m_protectedInlineeParsers;
-};
-
-WTF_MAKE_TZONE_ALLOCATED_IMPL(OMGIRGenerator);
 
 // Memory accesses in WebAssembly have unsigned 32-bit offsets, whereas they have signed 32-bit offsets in B3.
 int32_t OMGIRGenerator::fixupPointerPlusOffset(Value*& ptr, uint32_t offset)
@@ -7425,11 +6376,4 @@ auto OMGIRGenerator::addI64TruncUF32(ExpressionType argVar, ExpressionType& resu
     return { };
 }
 
-} } // namespace JSC::Wasm
-
-#include "WasmOMGIRGeneratorInlines.h"
-
 WTF_ALLOW_UNSAFE_BUFFER_USAGE_END
-
-#endif // USE(JSVALUE32_64)
-#endif // ENABLE(WEBASSEMBLY_OMGJIT)
